# Super fast XOR