Closes #76

```
benchmark                      old ns/op     new ns/op     delta
BenchmarkXOR/2/basic-2         5.43          5.46          +0.55%
BenchmarkXOR/2/fast-2          6.42          6.39          -0.47%
BenchmarkXOR/16/basic-2        17.0          17.1          +0.59%
BenchmarkXOR/16/fast-2         17.7          16.6          -6.21%
BenchmarkXOR/32/basic-2        30.6          30.7          +0.33%
BenchmarkXOR/32/fast-2         19.3          17.8          -7.77%
BenchmarkXOR/512/basic-2       448           445           -0.67%
BenchmarkXOR/512/fast-2        98.7          61.2          -37.99%
BenchmarkXOR/4096/basic-2      3487          3498          +0.32%
BenchmarkXOR/4096/fast-2       697           379           -45.62%
BenchmarkXOR/16384/basic-2     13894         13952         +0.42%
BenchmarkXOR/16384/fast-2      2738          1475          -46.13%

benchmark                      old MB/s     new MB/s     speedup
BenchmarkXOR/2/basic-2         368.43       365.98       0.99x
BenchmarkXOR/2/fast-2          311.55       313.15       1.01x
BenchmarkXOR/16/basic-2        941.21       935.48       0.99x
BenchmarkXOR/16/fast-2         901.42       962.22       1.07x
BenchmarkXOR/32/basic-2        1044.65      1041.27      1.00x
BenchmarkXOR/32/fast-2         1654.63      1802.62      1.09x
BenchmarkXOR/512/basic-2       1142.53      1149.90      1.01x
BenchmarkXOR/512/fast-2        5188.32      8368.92      1.61x
BenchmarkXOR/4096/basic-2      1174.33      1170.82      1.00x
BenchmarkXOR/4096/fast-2       5875.49      10802.77     1.84x
BenchmarkXOR/16384/basic-2     1179.14      1174.24      1.00x
BenchmarkXOR/16384/fast-2      5982.73      11104.28     1.86x
```