@nhooyr 2019-04-29
Need to ask why the for loops work but if statements after the 128 byte for loop slows things down to a crawl.